README.md
To run this app you must have the latest version of python installed. You can check this by running "python -v" in the terminal. You must have the AppOpener library installed to check this run
"pip install AppOpener in the terminal". Then, to run the app simply run "python app.py" in the terminal. This was created using flask.